修改注册代码

See merge request !9

+++ basic-user-service/src/main/java/com/lhiot/uc/basic/mapper/BaseUserMapper.java
@@ -13,6 +13,6 @@ public interface BaseUserMapper {
      * @param param
      * @return
      */
-    int save(BaseUser param);
+    Long save(BaseUser param);
 
 }
